% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficient method}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object from linreg class}

\item{...}{Additional arguments}
}
\description{
Coefficient method
}
