% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Coefficient method}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{An object from ridgereg class}

\item{...}{Additional arguments}
}
\description{
Coefficient method
}
