% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object from linreg class}

\item{...}{Additional arguments}
}
\description{
Plot method
}
