% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict method}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{ridgereg}.}

\item{newdata}{Optional new data frame for predictions. If NULL, returns fitted values.}

\item{...}{Additional arguments}
}
\value{
A numeric vector of predicted values.
}
\description{
Predict method
}
