% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An onject from linreg class}

\item{...}{Additional arguments}
}
\description{
Print method
}
