% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print method}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{An object from ridgereg class}

\item{...}{Additional arguments}
}
\description{
Print method
}
