% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary method}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object from linreg class}

\item{...}{Additional arguments}
}
\description{
Summary method
}
