% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Mean Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
A \code{ggplot} scatter plot representing mean departure delays.
}
\description{
This function produces a scatter plot that visualizes the mean departure delay
for flights departing from New York City airports. The plot is based on
aggregated flight delay data from the \code{flights} dataset in the
\code{nycflights13} package and airport information from the \code{airports}
data set in the same package.
}
\details{
The function uses \pkg{dplyr} for data transformation and \pkg{ggplot2} for
visualization. Airports are plotted according to their longitude and latitude,
with both color and point size representing the mean departure delay in minutes.

This function has no arguments and returns a ggplot object.
}
\examples{
visualize_airport_delays()

}
